pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "ihateai"
version = "0.0.1"
authors = [{ name = "Cody", email = "cody@dkdc.dev" }]
description = "ARC-AGI"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    # data stuff
    'ibis-framework[duckdb,deltalake]==9.1.0',
    # viz
    'plotly',
    'streamlit',
    # ml
    'mlflow',
    'optuna',
    'scikit-learn',
    'xgboost',
]

[project.urls]
"Homepage" = "https://github.com/lostmygithubaccount/ihateai"
"Bug Tracker" = "https://github.com/lostmygithubaccount/ihateai/issues"

[tool.ruff]
extend-include = ["*.ipynb"]
